requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
dataclasses-json==0.6.7
deepgram-sdk==4.8.1
deprecation==2.1.0
distro==1.9.0
elevenlabs==2.13.0
frozenlist==1.7.0
google-genai
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
marshmallow==3.26.1
multidict==6.6.4
mypy_extensions==1.1.0
packaging==25.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
pygame==2.6.1
python-dotenv==1.1.1
rhasspysilence
requests==2.32.5
sniffio==1.3.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1

